report/requirements_axes.txt
# Protection contre les attaques par force brute
django-axes==6.1.1
